Time: 29 ms (71.39%), Space: 21.7 MB (49.68%) - LeetHub

+++ 0907-koko-eating-bananas/0907-koko-eating-bananas.cpp
@@ -1,6 +1,6 @@
 class Solution {
 public:
-    int minEatingSpeed(vector<int>& piles, int h) {
+    long minEatingSpeed(vector<int>& piles, int h) {
 
         int maxi = *max_element(piles.begin(),piles.end());
